Create round_robin.c
Round robin is a scheduling algorithm used in computer operations that equally distributes work across all available resources. By doing this, it is made sure that no resource is overused, which may result in mistakes and other problems in the future. This is frequently described as a round robin scheduling method.